[sysmem] Remove zircon-migrated config from targets
zircon-migrated sets the ZX_ASSERT_LEVEL define, while in release builds
it's normally not set. That meant that device_test.cc was being compiled
with different options from buffer_collection.cc, so they disagreed
about the size of BufferCollection and that was causing
FakeDdkSysmem.NamedClient to sometimes read incorrect data and hang.

We don't need the zircon-migrated config for this code, so remove it.

Change-Id: I4467346d37fc293596a5e161af809cf502d1b36d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393064
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: John Bauman <jbauman@google.com>