[zxcrypt] Migrate to DdkUnbindNew.
To reply to unbind, UnbindTxn.Reply() is called rather than
DeviceRemoveDeprecated. Is is required to call Reply() or
save the UnbindTxn to be replied to later.

It is also guaranteed that DdkUnbindNew will only ever be called
once for a device added via DdkAdd. This means we can make the
following changes:

* Device::DdkUnbindNew
Since the deviceManager always calls Device::Init() before DdkAdd(),
and DdkUnbindNew is only called once, we do not need to check |active_|.

* DeviceManager::DdkUnbindNew
The device should not able to be in the kBinding state, as the state is
set to kSealed following DdkAdd while holding |mtx_|.
Since DdkUnbindNew is only called once for a successfully added device,
it should also not be in kUnbinding or kRemoved state.

TEST= runtests -t zxcrypt-test

BUG=37831

Change-Id: I821ea64c285c67c47689da2465b51ba25866c450
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397813
Reviewed-by: Drew Fisher <zarvox@google.com>
Testability-Review: Drew Fisher <zarvox@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>