[storage] Add the equivalent of RemoteBlockDevice for Rust.
Also fixed a bug in our storage driver that would drop responses if the
fifo became full. This isn't a problem with our C/C++ client because it
limits to a maximum of 8 in-flights, but the Rust implementation doesn't
have this limitation.

This does not support barriers or flush (yet).

Bug: 53796
Change-Id: Ib216c5e17539cc76e620081be25cc352b4a0fc6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399393
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Simon Shields <simonshields@google.com>