[dma-buffer] Fix memory leak
This patch fixes a memory leak in dma-buffer.
The RAII wrapper for PMT does not unpin the PMT,
so the memory was not being properly freed.
This patch manually invokes unpin to unpin the pages
so that the physical memory can be freed when the buffer
is destroyed.

Test: fx test dma-buffer-test
Change-Id: I9bac40c11e8f06cb11dc8b1fb7566e606bd5c5ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393634
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>