[as370] Migrate off DdkRemoveDeprecated().
Previously DdkRemoveDeprecated() was used to reply
to the unbind hook and also for the releasing of a device

UnbindTxn.Reply() is now used to reply to the unbind hook.

DdkAsyncRemove() is now used to schedule unbinding and releasing
of a device. Since Dwc2Device does not implement an unbind hook,
this does not change existing behavior.

BUG=37831

Change-Id: Ie3b8e897f4e8e874d39500d33600bb87ac6b3d89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394738
Reviewed-by: Carl Norum <cjn@google.com>
Testability-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>