[fragment][driver] Add SCPI protocol client
The fragment driver currently does not support the SCPI protocol. The
result is that fragments that are bound to a SCPI device end up with the
protocol ZX_PROTOCOL_MISC. This leads to a situation where a composite
device which contains SCPI fragment will be unable to access that
fragment.

Fixed: 56284
Test: run device-enumeration-test locally on VIM2
Change-Id: I6b0cff0c6bfe10bef9c1684105c2d098d7a3e9a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407278
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Testability-Review: Patrick Shickel <pshickel@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>