[nvme] Migrate from DEVICE_ADD_INVISIBLE to init hook.
The device init hook is called after device_add().

If the init hook is implemented, the device is guaranteed
to be invisible and unable to be unbound until
device_init_reply() is called.

If device_init_reply() is passed an error, the device
will be scheduled to be unbound.

BUG=43261

Change-Id: Ib6bd6228941fab5f575995e721348cd38406accb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406652
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>