[xhci-rewrite] Migrate from DEVICE_ADD_INVISIBLE to DdkInit.
If DdkInit is implemented, the device is guaranteed to be invisible
and not able to be unbound until InitTxn.Reply() is called. Replying
with an error will lead to the device being unbound.

The fake_ddk will automatically call DdkInit following DdkAdd. I have
moved the BufferFactory argument to the UsbXhci constructor, so the
core test can pass in the fake BufferFactory.

BUG=43261

Change-Id: I0ae643cfa5fa66adb38355238774651f1176a04f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408134
Reviewed-by: Brian Bosak <bbosak@google.com>
Testability-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>