[optee] Unpin shared memory pmt prior to bti close
The OP-TEE driver was letting its BTI handle drop immmediately after
pinning the shared memory region. This CL extends the lifetime of the
BTI handle to the lifetime of the device. The PMT is held by the
SharedMemoryManager as part of a ddk::MmioPinnedBuffer. It will be
unpinned when the SharedMemoryManager is destroyed. This CL also
implements the DdkSuspend hook for the controller so that all active
trusted app sessions are terminated and the shared memory is unpinned.

Fixed: 56233
Test: fx test optee-unittest
Change-Id: I679db0e9d2f530a7c537a8ae70b28aa30fd56553
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407213
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: John Grossman <johngro@google.com>