[fvm] Migrate off DEVICE_ADD_INVISIBLE, DdkRemoveDeprecated.
The DdkInit hook replaces DEVICE_ADD_INVISIBLE. If implemented,
it is called after DdkAdd(), and it is guaranteed the device
will be invisible and not unbound until InitTxn.Reply().
Replying with an error will schedule the device to be unbound.

DdkAsyncRemove() replaces DdkRemoveDeprecated(). This schedules
the unbind hook to be called (and eventually the release hook)
and is safe to call multiple times.

---------------
DDK lifecycle:
---------------
1) DdkAdd()
2) DdkInit()
3) InitTxn.Reply() // Device becomes visible
4) DdkUnbind()     // started by InitTxn.Reply(!ZX_OK) / DdkAsyncRemove
5) UnbindTxn.Reply()
6) DdkRelease()

TEST=  fx test fvm-driver-integration-test

BUG=37831
BUG=43261

Change-Id: I980cdda2855b45fc3beb4271d2c35a87dec47f70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407130
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>