[usb][hub] validate descriptor returned bLength
The hub driver requests a number of descriptors from the hub device.
The current implementation is not verifying that the length, as reported
by the descriptors, equals the total returned descriptor bytes.  If this
happens, it might be a malformed descriptor which has the potential to
leak out-of-bounds memory.

This adds .bLength (and .bDescLength) bounds checking.

Bug: 57060
Bug: 57059
Change-Id: I8374b6f107efe8d48b7f1af19e834986e07f540f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410903
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>