[usb-hub-rewrite] Initialize usb_ and bus_ prior to calling DdkAdd
In the normal DDK, DdkInit is invoked asynchronously. However,
in fake_ddk, DdkInit is invoked from within the call to DdkAdd.
This allows for the async loop to be initialized prior to Init completing,
resulting in a race when attempting to access bus_ or usb_.

Test: UBSAN in infra
Change-Id: Idf66895afe9995261f3a1bfccf1ad8e54beb48a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413082
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>