[acpi][x64] Break initial namespace walk into 2 passes.
Previously to this CL, the code which was walking the ACPI device tree
would look for an HDAS (Intel HDA audio controller device) device and
attempt to publish the NHLT as metadata bound to the controller's PCI
topological path for the device to pick up later on after the PCI bus
driver had published the controller's device.

The trouble here is that the PCI bus device was always going to be
encountered first, before the HDAS device (which is a child of the PCI
bus device).  This meant that the PCI bus device was being published,
allowing the driver to start initializing before the ACPI driver had a
chance to publish the HDAS NHLT metadata, and setting up a race to see
if the ACPI driver could publish the NHLT metadata before the HDA
driver loaded and went looking for it.

Now, the init phase has been broken into two passes.  During the first
pass, we handle any device quirks and publish any metadata.  During
the second pass, we actually start to publish devices

Change-Id: I50ee509402577acc20d1c655fb5ca7b315b6504d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411941
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>