[aml-power][s905d2] Introduce power driver.
This patch introduces an power implementation driver for the amlogic
platform. Currently it is only implemented for the s905d2's CPU
power domain and currently only supports changing the voltage of the
power domain.

The functionality is duplicated from the aml-voltage component of the
aml-thermal driver which is ultimately planned to be removed from the
aml-thermal driver.

Change-Id: I25035d3ea41c1f75b27fba82631f75064b6c0ee5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401380
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>