[fidl][llcpp] merge Reply and ReplyWithStatus
Now Reply returns a status (which can be ignored).

Fixes the code which use it.

Also remove the version with a DecodedMessage argument which is not used
and not tested. This version can be reintroduced later with another type
(DecodedMessage will be removed).

That spares 75484 bytes.

Change-Id: I6b3a7874004ce1aba439333a3be4f26da58218a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413020
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Madhav Iyengar <madhaviyengar@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>