[audio][astro] Move enable gpio to controller
Having the control of the enable gpio in the controller facilitates
configurations where more than one codec is connected to the enable
gpio, as in the case in sherlock. With this we consider the enable
gpio an input to the codec and not in control of the codec.

Note that the board driver now hardware resets the codec via the GPIO
on initialization.

Test: tas27xx-test
Test: astro-tdm-test
Test: dvice-enumeration-test on Astro
Change-Id: I3e8b4b83c135b8fc5a9fafdf615986241cccfeea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413878
Reviewed-by: Carl Norum <cjn@google.com>
Reviewed-by: Eric Holland <hollande@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>