[audio][astro] Fix channels configuration
We allow masking of channels as they are provided to the
driver in the ring buffer. The enablement in the AMLogic
SoC was incorrectly using the second channel (Left) which
would be filtered out if the masks for channels to enable
only had setup the first (Right) channel. This worked when
no filtering was applied in the driver (as in tones via
audio-driver-ctl which does it in software), but failed
when filtering is applied in the driver (as in
signal_generator and audio_core). The fix is to configure
the correct (right) channel as the only mono channel used
and make sure the codec also uses the right channel (asserts
added).

Fixed: 58173

Test: tas27xx-test
Test: Manual check of sounds from button pressing
Test: signal_generator --sine=440 --amp=0.95 --dur=1
Change-Id: I72cacea8e9b77c9f4f62f719ecf8e7a724468406
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416322
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
Reviewed-by: Eric Holland <hollande@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>