[camera][imx355] Connect the driver to the board
This CL establishes the connections between the hardware resources,
such as GPIOs and I2C interfaces, to the IMX355 camera sensor driver.
It also adds the imx355 to the device enumeration test.

Note that the EEPROM in the IMX355 camera modules will be connected
in a future CL.

Change-Id: I9eba0e43bca8b4e066d3330a7480d3be8495f5e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417623
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: John Sasinowski <jsasinowski@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Nic Zoghb <nzo@google.com>