[thermal][luis] Support aml-thermal-s905d2g
This change supports the migration from aml-thermal-s905d2g-legacy to
aml-thermal-s905d2g on Luis. It is the first change in a series to
achieve a soft migration:
  1. Add thermal devices to support both drivers on Luis
  2. Stop building aml-thermal-s905d2g-legacy on Luis; instead build
     aml-thermal-s905d2g
  3. Remove the thermal device that supports aml-thermal-s905d2g-legacy
     on Luis

With this change, both a platform device (to support
aml-thermal-s905d2g) and a composite device (to support
aml-thermal-s905d2g-legacy) will be added. The driver that ultimately
gets matched and loaded will be determined by the build configuration
(currently it is aml-thermal-s905d2g-legacy, but the next change in this
series will update it to be aml-thermal-s905d2g). Only one of the two
drivers should be built into the image for this reason.

Test: boot up Luis board with this change and verify power_manager is
  still successfully reading from the thermal driver
Change-Id: Ib80efdc09b6f43b1edc7aad972345446ad8ff549
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420743
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Testability-Review: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>