[fidl][llcpp] optimize sync event handler
The callbacks for handling events are modified. Now we only give a
pointer to the primary object. This way we avoid unnecessary moves.

HandleEvents now returns a fidl::Result

HandleEvents now uses a reference to EventHandlers and uses
fit::function instead of fit::callback. That means that users can now
create the handlers only once and use them for each invocation.

SyncClient::HandleEvents is now inlined.

That reduces the size of the generated code. That also reduces the size
of the user code.

size reduction of LLCPP Domain Objects: 6836 bytes
global size reduction (LLCPP + user code): 16384 bytes

The goal for this CL and the upcomming CLs is to only have callbacks
with one argument (a pointer to the primary object). This way, all the
callbacks will be consistent and easier to use. Most of all, by removing
all the moves (for the arguments), we are able to reduce the code size.

Change-Id: I2a0e29c85de5d89703b85590a0e761f4917bebfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421236
Commit-Queue: Vincent Belliard <vbelliard@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>