[sy8827][luis] Pass i2c params to sy8827 bucks.
There are two of these on the system and we need to be
able to distinguish between them somehow.

Pass the i2c address of the bucks as an argument so that they
can set it as a device property when they add a device.

Change-Id: I994f356078efd69212fec03d45e10cf84ab1ba94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421800
Reviewed-by: Carl Norum <cjn@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Andres Oportus <andresoportus@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>