[storage] Fix shutdown deadlock.
Driver destructors and Unbind calls are all processed from the main
thread and care must be taken not to block that thread.

Consider the following scenario:

  ram-nand > ftl > fvm

If you called unbind on the ram-nand and ftl drivers, it would try and
terminate their worker threads.  Later, the destructors for those
drivers are called.  Prior to this fix, there could be unprocessed
messages in the ram-nand and ftl buffers that wouldn't have been cleaned
up until the destructor is called.  The destructor for the ftl driver
would wait until all outstanding messages have been responded to (it has
to because otherwise those callbacks could come later which would lead
to a use-after-free issue).  So, it was possible to end up with a
deadlock where the destructor for the ftl driver would be waiting for
responses to all outstanding messages, but that wouldn't happen until
the destructor for the ram-nand driver had run, but that can't run
because it's single threaded and couldn't until the ftl's destructor had
finished.

The fix is to flush unresponded messages in the worker threads.

This might a problem with other storage drivers, but these are the two
cases I had issues with on a test I'm working on.

Bug: 59224
Change-Id: I7ccb0765a276a97138d1dd71d8db349d9988d08a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421478
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>