[fidl] Specialize IsFoo<> LLCPP traits using codegen
Previously, these traits were implemented based on the
attributes of the classes themselves (e.g. checking for
a ::Type, ::MaxNumHandles, or a builder method), which
is not robust to changes to the API. Instead, this CL
emits specializations for each type in the generated code.

Test: fx build fidl_llcpp_types_test
Change-Id: Ic32c14e3a1dcd4b9f98a1a3f6cea473e65e300bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423115
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>